+++ source/projects/chugins/ConvRev/ConvRev.cpp
@@ -23,7 +23,6 @@
 #include <iostream>
 #include <mutex>
 #include <string>
-
 #include <thread>
 #include <vector>